backend/trigger-engine/requirements.txt
celery==5.3.6
redis==5.0.3
requests==2.31.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
schedule==1.2.1
